+++ backend/springboot/db/mozara.sql
@@ -70,36 +70,8 @@ CREATE TABLE `daily_habits` (
 
 LOCK TABLES `daily_habits` WRITE;
 /*!40000 ALTER TABLE `daily_habits` DISABLE KEYS */;
-INSERT INTO `daily_habits` VALUES (1,'매일 물 7잔 이상 마시기','물 마시기',10,'routine'),(2,'이펙터 4번 사용','이펙터 사용',10,'routine'),(3,'탈모 에센스로 직접 개선','아침 부스터 사용',5,'routine'),(4,'탈모 에센스로 직접 개선','밤 부스터 사용',5,'routine'),(5,'상열감 감소로 탈모 예방','백회혈/사신총혈 마사지',5,'routine'),(6,'혈액 순환 촉진 및 염증 완화','오메가-3 섭취',5,'nutrient '),(7,'모낭 자극 및 모발 성장 촉진','비타민 D 섭취',5,'nutrient '),(8,'항산화 작용 및 건조함 완화','비타민 E 섭취',5,'nutrient '),(9,'모발 구성 성분 및 성장 촉진','단백질 섭취',5,'nutrient '),(10,'산소 운반 및 모발 건강 유지','철분 섭취',5,'nutrient '),(11,'모발 성장 및 강화 촉진','비오틴 섭취',5,'nutrient '),(12,'모발 성장 및 재생 촉진','아연 섭취',5,'nutrient '),(13,'모공 청결 유지로 탈모 방지','밤에 머리 감기',5,'cleanliness'),(14,'모발 약화 및 냉기 방지','머리 바싹 말리기',5,'cleanliness'),(15,'상열감 감소로 탈모 예방','백회혈/사신총혈 마사지',5,'cleanliness'),(16,'머리 엉킴 방지 및 노폐물 제거','샴푸 전 머리 빗질',5,'cleanliness'),(17,'두피 영양 공급 및 보습','두피 영양팩하기',5,'cleanliness');
+INSERT INTO `daily_habits` VALUES (1,'매일 물 7잔 이상 마시기','물 마시기',10,'routine'),(2,'이펙터 4번 사용','이펙터 사용',10,'routine'),(3,'탈모 에센스로 직접 개선','아침 부스터 사용',5,'routine'),(4,'탈모 에센스로 직접 개선','밤 부스터 사용',5,'routine'),(5,'상열감 감소로 탈모 예방','백회혈/사신총혈 마사지',5,'routine'),(6,'혈액 순환 촉진 및 염증 완화','오메가-3 섭취',5,'nutrient'),(7,'모낭 자극 및 모발 성장 촉진','비타민 D 섭취',5,'nutrient'),(8,'항산화 작용 및 건조함 완화','비타민 E 섭취',5,'nutrient'),(9,'모발 구성 성분 및 성장 촉진','단백질 섭취',5,'nutrient'),(10,'산소 운반 및 모발 건강 유지','철분 섭취',5,'nutrient'),(11,'모발 성장 및 강화 촉진','비오틴 섭취',5,'nutrient'),(12,'모발 성장 및 재생 촉진','아연 섭취',5,'nutrient'),(13,'모공 청결 유지로 탈모 방지','밤에 머리 감기',5,'cleanliness'),(14,'모발 약화 및 냉기 방지','머리 바싹 말리기',5,'cleanliness'),(15,'머리 엉킴 방지 및 노폐물 제거','샴푸 전 머리 빗질',5,'cleanliness'),(16,'두피 영양 공급 및 보습','두피 영양팩하기',5,'cleanliness');
 /*!40000 ALTER TABLE `daily_habits` ENABLE KEYS */;
-UNLOCK TABLES;
-
---
--- Table structure for table `measurement_logs`
---
-
-DROP TABLE IF EXISTS `measurement_logs`;
-/*!40101 SET @saved_cs_client     = @@character_set_client */;
-/*!50503 SET character_set_client = utf8mb4 */;
-CREATE TABLE `measurement_logs` (
-  `log_id` int NOT NULL AUTO_INCREMENT,
-  `type` varchar(255) DEFAULT NULL,
-  `value` varchar(255) DEFAULT NULL,
-  `record_date` datetime DEFAULT NULL,
-  `user_id_foreign` int DEFAULT NULL,
-  PRIMARY KEY (`log_id`),
-  KEY `user_id_foreign` (`user_id_foreign`),
-  CONSTRAINT `measurement_logs_ibfk_1` FOREIGN KEY (`user_id_foreign`) REFERENCES `users` (`user_id`)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
-/*!40101 SET character_set_client = @saved_cs_client */;
-
---
--- Dumping data for table `measurement_logs`
---
-
-LOCK TABLES `measurement_logs` WRITE;
-/*!40000 ALTER TABLE `measurement_logs` DISABLE KEYS */;
-/*!40000 ALTER TABLE `measurement_logs` ENABLE KEYS */;
 UNLOCK TABLES;
 
 --
@@ -137,11 +109,11 @@ DROP TABLE IF EXISTS `user_habit_log`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `user_habit_log` (
-  `log_id` int NOT NULL AUTO_INCREMENT,
+  `habit_log_id` int NOT NULL AUTO_INCREMENT,
   `habit_id_foreign` int DEFAULT NULL,
   `user_id_foreign` int DEFAULT NULL,
   `completion_date` date DEFAULT NULL,
-  PRIMARY KEY (`log_id`),
+  PRIMARY KEY (`habit_log_id`),
   KEY `habit_id_foreign` (`habit_id_foreign`),
   KEY `user_id_foreign` (`user_id_foreign`),
   CONSTRAINT `user_habit_log_ibfk_1` FOREIGN KEY (`habit_id_foreign`) REFERENCES `daily_habits` (`habit_id`),
@@ -184,6 +156,34 @@ CREATE TABLE `user_log` (
 LOCK TABLES `user_log` WRITE;
 /*!40000 ALTER TABLE `user_log` DISABLE KEYS */;
 /*!40000 ALTER TABLE `user_log` ENABLE KEYS */;
+UNLOCK TABLES;
+
+--
+-- Table structure for table `user_metrics`
+--
+
+DROP TABLE IF EXISTS `user_metrics`;
+/*!40101 SET @saved_cs_client     = @@character_set_client */;
+/*!50503 SET character_set_client = utf8mb4 */;
+CREATE TABLE `user_metrics` (
+  `user_metrics_id` int NOT NULL AUTO_INCREMENT,
+  `type` varchar(255) DEFAULT NULL,
+  `value` varchar(255) DEFAULT NULL,
+  `record_date` datetime DEFAULT NULL,
+  `user_id_foreign` int DEFAULT NULL,
+  PRIMARY KEY (`user_metrics_id`),
+  KEY `user_id_foreign` (`user_id_foreign`),
+  CONSTRAINT `user_metrics_ibfk_1` FOREIGN KEY (`user_id_foreign`) REFERENCES `users` (`user_id`)
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
+/*!40101 SET character_set_client = @saved_cs_client */;
+
+--
+-- Dumping data for table `user_metrics`
+--
+
+LOCK TABLES `user_metrics` WRITE;
+/*!40000 ALTER TABLE `user_metrics` DISABLE KEYS */;
+/*!40000 ALTER TABLE `user_metrics` ENABLE KEYS */;
 UNLOCK TABLES;
 
 --
@@ -253,4 +253,4 @@ UNLOCK TABLES;
 /*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
 /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 
--- Dump completed on 2025-09-16 14:45:14
+-- Dump completed on 2025-09-17 10:48:07
