+++ backend/springboot/src/main/java/com/example/springboot/service/ai/PythonBackendHealthService.java
@@ -1,0 +1,33 @@
+package com.example.springboot.service.ai;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.*;
+import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class PythonBackendHealthService {
+
+    @Value("${ai.python.base-url:http://localhost:8000}")
+    private String pythonBaseUrl;
+
+    private final RestTemplate restTemplate = new RestTemplate();
+
+    /**
+     * Python 백엔드 연결 상태 확인
+     */
+    public boolean checkPythonBackendHealth() {
+        try {
+            String url = pythonBaseUrl + "/";
+            ResponseEntity<String> response = restTemplate.getForEntity(url, String.class);
+            return response.getStatusCode() == HttpStatus.OK;
+        } catch (Exception e) {
+            log.warn("Python 백엔드 연결 확인 실패: {}", e.getMessage());
+            return false;
+        }
+    }
+}
